# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Education {
  degree: string;
  institution: string;
  year: string;
}

export interface Experience {
  role: string;
  organization: string;
  dates: string;
}

export type ProfessionalStatus = typeof ProfessionalStatus[keyof typeof ProfessionalStatus];


export const ProfessionalStatus = {
  open: 'open',
  selective: 'selective',
  unavailable: 'unavailable',
} as const;

export interface Professional {
  id: string;
  name: string;
  slug: string;
  profession: string;
  location: string;
  initials: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  companyLogoUrl?: string | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  completion: number;
  status: ProfessionalStatus;
  summary: string;
  skills: string[];
  education: Education[];
  experience: Experience[];
  interests?: string[];
  qrValue: string;
  createdAt: string;
}

export interface ProfessionalInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  profession: string;
  /** @minLength 2 */
  location: string;
  company?: string;
  summary?: string;
  imageUrl?: string;
  companyLogoUrl?: string;
  skills?: string[];
  education?: Education[];
  experience?: Experience[];
  interests?: string[];
}

export type ProfessionalUpdateStatus = typeof ProfessionalUpdateStatus[keyof typeof ProfessionalUpdateStatus];


export const ProfessionalUpdateStatus = {
  open: 'open',
  selective: 'selective',
  unavailable: 'unavailable',
} as const;

export interface ProfessionalUpdate {
  /** @minLength 2 */
  name?: string;
  /** @minLength 2 */
  profession?: string;
  /** @minLength 2 */
  location?: string;
  company?: string;
  summary?: string;
  imageUrl?: string;
  companyLogoUrl?: string;
  skills?: string[];
  education?: Education[];
  experience?: Experience[];
  interests?: string[];
  status?: ProfessionalUpdateStatus;
}

export interface Activity {
  id: string;
  label: string;
  detail: string;
  date: string;
}