# lib/api-zod/src/generated/types/professional.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Education } from './education';
import type { Experience } from './experience';
import type { ProfessionalStatus } from './professionalStatus';

export interface Professional {
  id: string;
  name: string;
  slug: string;
  profession: string;
  location: string;
  initials: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  companyLogoUrl?: string | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  completion: number;
  status: ProfessionalStatus;
  summary: string;
  skills: string[];
  education: Education[];
  experience: Experience[];
  interests?: string[];
  qrValue: string;
  createdAt: string;
}