# lib/api-zod/src/generated/types/professionalInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Education } from './education';
import type { Experience } from './experience';

export interface ProfessionalInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  profession: string;
  /** @minLength 2 */
  location: string;
  company?: string;
  summary?: string;
  imageUrl?: string;
  companyLogoUrl?: string;
  skills?: string[];
  education?: Education[];
  experience?: Experience[];
  interests?: string[];
}